+++ lib/montague/model/journals_report.rb
@@ -3,15 +3,11 @@ module Montague
     # Journals report
     #
     class JournalsReport
-      include Montague::Model::ReportHTTPResponseMixin
       include Montague::Model::ReportHeaderMixin
+      include Montague::Model::ReportHTTPResponseMixin
 
       # @return [Array<Montague::Model::Journal>]
       attr_accessor :journals
-
-      # Present only when #journals size is 1
-      # @return [Montague::Model::Publisher, nil]
-      attr_accessor :publisher
 
     end
   end