+++ lib/montague/model/report_publisher_mixin.rb
@@ -1,0 +1,12 @@
+module Montague
+  module Model
+    # Report publisher mixin
+    #
+    module ReportPublisherMixin
+
+      # @return [Montague::Model::Publisher]
+      attr_accessor :publisher
+
+    end
+  end
+end