+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.1.0 - 2018-01-26
+## 0.1.0 - 2018-02-05
 ### Added
 - For SHERPA/RoMEO API v2.9.
-- Find journal by ISSN, title.
-- Find journals by partial title.
+- Find publisher indirectly using journal search by ISSN, search by title.
+- Find publisher directly using publisher search by name, search by ID.